src/requirements.txt
Django~=4.0.6
django-environ~=0.9.0
psycopg2~=2.9.3
djangorestframework~=3.13.1
gunicorn~=20.1.0
django-filter~=22.1
django-money~=3.0.0
requests~=2.25.1
djoser~=2.1.0
pycryptodome~=3.15.0